+++ examples/simple/foo/views.py
@@ -1,19 +1,22 @@
 from django.http import Http404
-from django.shortcuts import render_to_response
+from django.shortcuts import render
 from django.template import RequestContext
 from django.views.generic.base import View
 from django.core.urlresolvers import reverse
 
 from ska import sign_url
 from ska.defaults import DEFAULT_PROVIDER_PARAM
-from ska.contrib.django.ska.decorators import validate_signed_request, m_validate_signed_request
+from ska.contrib.django.ska.decorators import (
+    validate_signed_request,
+    m_validate_signed_request
+)
 from ska.contrib.django.ska.settings import SECRET_KEY, PROVIDERS
 
 from foo.models import FooItem
 
+
 def browse(request, template_name='foo/browse.html'):
-    """
-    Foo items listing.
+    """Foo items listing.
 
     :param django.http.HttpRequest request:
     :param str template_name:
@@ -21,15 +24,13 @@ def browse(request, template_name='foo/browse.html'):
     """
     queryset = FooItem._default_manager.all()
 
-    context = {
-        'items': queryset,
-    }
+    context = {'items': queryset,}
 
-    return render_to_response(template_name, context, context_instance=RequestContext(request))
+    return render(request, template_name, context)
+
 
 def authenticate(request, template_name='foo/authenticate.html'):
-    """
-    Authenticate.
+    """Authenticate.
 
     :param django.http.HttpRequest request:
     :param str template_name:
@@ -42,32 +43,38 @@ def authenticate(request, template_name='foo/authenticate.html'):
     remote_ska_login_urls = []
     for i in range(3):
         signed_remote_ska_login_url = sign_url(
-            auth_user = 'test_ska_user_{0}'.format(i),
-            secret_key = SECRET_KEY, # Using general secret key
-            url = remote_ska_login_url,
-            extra = {
+            auth_user='test_ska_user_{0}'.format(i),
+            # Using general secret key
+            secret_key=SECRET_KEY,
+            url=remote_ska_login_url,
+            extra={
                 'email': 'test_ska_user_{0}@mail.example.com'.format(i),
                 'first_name': 'John {0}'.format(i),
                 'last_name': 'Doe {0}'.format(i),
             }
-            )
-        remote_ska_login_urls.append((i, remote_ska_login_url, signed_remote_ska_login_url))
+        )
+        remote_ska_login_urls.append(
+            (i, remote_ska_login_url, signed_remote_ska_login_url)
+        )
 
     # Login URLs by provider
     remote_ska_login_urls_by_provider = []
     for uid, data in PROVIDERS.items():
         signed_remote_ska_login_url = sign_url(
-            auth_user = 'test_ska_user_{0}'.format(uid),
-            secret_key = data.get('SECRET_KEY'), # Using provider-specific secret key
-            url = remote_ska_login_url,
-            extra = {
+            auth_user='test_ska_user_{0}'.format(uid),
+            # Using provider-specific secret key
+            secret_key=data.get('SECRET_KEY'),
+            url=remote_ska_login_url,
+            extra={
                 'email': 'test_ska_user_{0}@mail.example.com'.format(uid),
                 'first_name': 'John {0}'.format(uid),
                 'last_name': 'Doe {0}'.format(uid),
                 DEFAULT_PROVIDER_PARAM: uid,
             }
-            )
-        remote_ska_login_urls_by_provider.append((uid, remote_ska_login_url, signed_remote_ska_login_url))
+        )
+        remote_ska_login_urls_by_provider.append(
+            (uid, remote_ska_login_url, signed_remote_ska_login_url)
+        )
 
     # Template context
     context = {
@@ -75,7 +82,8 @@ def authenticate(request, template_name='foo/authenticate.html'):
         'remote_ska_login_urls_by_provider': remote_ska_login_urls_by_provider
     }
 
-    return render_to_response(template_name, context, context_instance=RequestContext(request))
+    return render(request, template_name, context)
+
 
 @validate_signed_request()
 def detail(request, slug, template_name='foo/detail.html'):
@@ -94,12 +102,12 @@ def detail(request, slug, template_name='foo/detail.html'):
 
     context = {'item': item}
 
-    return render_to_response(template_name, context, context_instance=RequestContext(request))
+    return render(request, template_name, context)
+
 
 class FooDetailView(View):
-    """
-    Class based view.
-    """
+    """Class based view."""
+
     @m_validate_signed_request()
     def get(self, request, slug, template_name='foo/detail.html'):
         try:
@@ -109,15 +117,15 @@ class FooDetailView(View):
 
         context = {'item': item}
 
-        return render_to_response(template_name, context, context_instance=RequestContext(request))
+        return render(request, template_name, context)
+
 
 def logged_in(request, template_name='foo/logged_in.html'):
-    """
-    Logged in landing page.
+    """Logged in landing page.
 
     :param django.http.HttpRequest request:
     :param str template_name:
     :return django.http.HttpResponse:
     """
     context = {}
-    return render_to_response(template_name, context, context_instance=RequestContext(request))
+    return render(request, template_name, context)