+++ scripts/test.sh
@@ -5,5 +5,5 @@ reset
 python src/ska/tests.py
 
 # django-ska tests
-python example/example/manage.py migrate --noinput --traceback -v 3
-python example/example/manage.py test ska --traceback -v 3
+python examples/simple/manage.py migrate --noinput --traceback -v 3
+python examples/simple/manage.py test ska --traceback -v 3