+++ src/ska/contrib/django/ska/apps.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 __title__ = 'ska.contrib.django.ska.apps'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2013-2016 Artur Barseghyan'
@@ -11,6 +13,7 @@ try:
         """Config."""
 
         name = 'ska.contrib.django.ska'
+        # label = 'ska_contrib_django_ska'
         label = 'ska'
 
 except ImportError: