+++ src/ska/contrib/django/ska/exceptions.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 # __title__ = 'ska.contrib.django.ska.exceptions'
 # __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 # __copyright__ = '2013-2016 Artur Barseghyan'