+++ src/ska/contrib/django/ska/tests/__init__.py
@@ -1,1 +1,3 @@
+from __future__ import absolute_import
+
 from .test_django_ska import *